Successfully saving to local server

+++ todo-list-app/src/app/views/created-task-lists-view/created-task-lists-view.component.ts
@@ -88,7 +88,12 @@ export class CreatedTaskListsViewComponent {
   }
 
   public onDeleteTaskList(list: TaskList) {
-    this._listManager.deleteList(list);
+    const prompt = 'Are you sure you want to delete this list?' +
+                   '\nAll of the tasks will also be deleted.';
+
+    if (confirm(prompt)) {
+      this._listManager.deleteList(list);
+    }
   }
 
   // private validateName(control: FormControl) {
